Move TF-specific files to tf/ folder
