+++ ml-agents/mlagents/trainers/components/reward_signals/curiosity/model.py
@@ -1,7 +1,7 @@
 from typing import List, Tuple
 from mlagents.tf_utils import tf
 
-from mlagents.trainers.models import ModelUtils
+from mlagents.trainers.tf.models import ModelUtils
 from mlagents.trainers.policy.tf_policy import TFPolicy
