+++ ml-agents/mlagents/trainers/sac/network.py
@@ -1,6 +1,6 @@
 from typing import Dict, Optional
 from mlagents.tf_utils import tf
-from mlagents.trainers.models import ModelUtils, EncoderType
+from mlagents.trainers.tf.models import ModelUtils, EncoderType
 
 LOG_STD_MAX = 2
 LOG_STD_MIN = -20
