+++ ml-agents/mlagents/trainers/tests/test_nn_policy.py
@@ -9,7 +9,7 @@ from mlagents.tf_utils import tf
 
 
 from mlagents.trainers.policy.tf_policy import TFPolicy
-from mlagents.trainers.models import EncoderType, ModelUtils, Tensor3DShape
+from mlagents.trainers.tf.models import EncoderType, ModelUtils, Tensor3DShape
 from mlagents.trainers.exception import UnityTrainerException
 from mlagents.trainers.tests import mock_brain as mb
 from mlagents.trainers.settings import TrainerSettings, NetworkSettings
