+++ ml-agents/mlagents/trainers/tests/test_simple_rl.py
@@ -27,7 +27,7 @@ from mlagents.trainers.settings import (
     RewardSignalType,
 )
 from mlagents.trainers.environment_parameter_manager import EnvironmentParameterManager
-from mlagents.trainers.models import EncoderType, ScheduleType
+from mlagents.trainers.tf.models import EncoderType, ScheduleType
 from mlagents_envs.side_channel.environment_parameters_channel import (
     EnvironmentParametersChannel,
 )
